Improve BlogList aesthetics and update avatars
Updated some of the blog posts to remove the title and date from the
actual content. Switched around BlogList to put the list of blog posts up
top and put the actual content below that. Also changed the current blog
header to read and display the blog title (which comes from
blog_content.json)

Also updated some of the avatars to be smaller. Added a small version of
avatar1 to display.